+++ app/src/main/AndroidManifest.xml
@@ -47,13 +47,6 @@
         android:theme="@style/Theme.AutoLightAndroid"
         tools:targetApi="31">
         <activity
-            android:name=".BluetoothActivity"
-            android:exported="false">
-            <meta-data
-                android:name="android.app.lib_name"
-                android:value="" />
-        </activity>
-        <activity
             android:name=".ControlLight"
             android:screenOrientation="portrait"
             android:configChanges="keyboardHidden|orientation"
