--- app/src/main/res/layout/activity_bluetooth.xml
@@ -1,94 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="#F1F1F1"
-    tools:context=".BluetoothActivity">
-
-    <ImageButton
-        android:id="@+id/paired_device"
-        android:layout_width="360dp"
-        android:layout_height="52dp"
-        android:layout_marginTop="88dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.49"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:srcCompat="@drawable/main_button_bluetooth" />
-
-    <ImageButton
-        android:id="@+id/search_device"
-        android:layout_width="360dp"
-        android:layout_height="52dp"
-        android:layout_marginTop="28dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.49"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/paired_device"
-        app:srcCompat="@drawable/main_button_bluetooth" />
-
-    <ImageButton
-        android:id="@+id/back_button"
-        android:layout_width="48dp"
-        android:layout_height="48dp"
-        android:layout_marginStart="16dp"
-        android:layout_marginTop="16dp"
-        android:src="@drawable/back_button"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <TextView
-        android:id="@+id/title"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="28dp"
-        android:text="블루투스 연결"
-        android:textColor="#3E3E3E"
-        android:textSize="20sp"
-        android:textStyle="bold"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.498"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <ListView
-        android:id="@+id/device_list"
-        android:layout_width="360dp"
-        android:layout_height="380dp"
-        android:layout_marginTop="28dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.49"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/search_device" />
-
-    <TextView
-        android:id="@+id/textView"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="페어링 목록"
-        android:textColor="#3E3E3E"
-        android:textSize="18sp"
-        app:layout_constraintBottom_toBottomOf="@+id/paired_device"
-        app:layout_constraintEnd_toEndOf="@+id/paired_device"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toStartOf="@+id/paired_device"
-        app:layout_constraintTop_toTopOf="@+id/paired_device"
-        app:layout_constraintVertical_bias="0.5" />
-
-    <TextView
-        android:id="@+id/textView2"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="기기 검색"
-        android:textColor="#3E3E3E"
-        android:textSize="18sp"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintVertical_bias="0.5"
-        app:layout_constraintBottom_toBottomOf="@+id/search_device"
-        app:layout_constraintEnd_toEndOf="@+id/search_device"
-        app:layout_constraintStart_toStartOf="@+id/search_device"
-        app:layout_constraintTop_toTopOf="@+id/search_device" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
