+++ app/src/main/res/layout/activity_control_light.xml
@@ -37,6 +37,7 @@
         android:layout_height="wrap_content"
         android:layout_marginTop="30dp"
         android:text="밝기 조정"
+        android:textColor="#3E3E3E"
         android:textSize="20sp"
         android:textStyle="bold"
         opencv:layout_constraintEnd_toEndOf="parent"
@@ -50,6 +51,7 @@
         android:layout_height="wrap_content"
         android:layout_marginTop="96dp"
         android:gravity="center"
+        android:textColor="#575757"
         android:text="얼굴을 잘 인식할 수 있도록\n얼굴을 가리는 물건(마스크, 선글라스 등)\n은 최대한 피해주세요."
         android:textSize="16sp"
         opencv:layout_constraintEnd_toEndOf="parent"
